PP-14728 - ParityCheckerResource put back and moved the test to correct folder and openapi-config.yaml regenerated
