+++ src/main/resources/openapi-config.yaml
@@ -15,7 +15,7 @@ resourceClasses:
   - uk.gov.pay.connector.paymentprocessor.resource.CardResource
   - uk.gov.pay.connector.paymentprocessor.resource.DiscrepancyResource
   - uk.gov.pay.connector.refund.resource.RefundsResource
-  - uk.gov.pay.connector.report.resource.ParityCheckerResource
+  - uk.gov.pay.connector.tasks.resource
   - uk.gov.pay.connector.token.resource.SecurityTokensResource
   - uk.gov.pay.connector.usernotification.resource.EmailNotificationResource
   - uk.gov.pay.connector.webhook.resource.NotificationResource
